LibOpenCV.jl

LibOpenCV — Module.

A special package that manages opencv binary depedencies

Pkg.build("LibOpenCV")

try to search your system opencv libraries and its dependencies, and throws errors if any issues. If opencv librarires are not found, it will install fresh opencv libraries into deps directrory, but not recommended unless if you have perfect requiremsnts to build opencv.

source

Index

LibOpenCV.find_library_e

Reference

LibOpenCV.find_library_e — Function.

find_library_e(mod)
find_library_e(mod, libdirs)
find_library_e(mod, libdirs, ext)

It tries to search the specified library by name. Not exported, but meant to be used by other opencv packages.

Parameters

mod : Module name

libdirs : library seach directries (default is dir of libopencv_highgui)

ext : library extention name (e.g. .so)

Retures

libpath : library path if found, othrewise return C_NULL

Examples

From the CVHighGUI.jl package,

libopencv_highgui = LibOpenCV.find_library_e("libopencv_highgui")
try
    Libdl.dlopen(libopencv_highgui, Libdl.RTLD_GLOBAL)
catch e
    warn("You might need to set DYLD_LIBRARY_PATH to load dependencies proeprty.")
    rethrow(e)
end

source